requirements.txt
aniso8601==8.1.0
appdirs==1.4.4
arrow==0.17.0
atomicwrites==1.4.0
attrs==20.3.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
black==20.8b1
blinker==1.4
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
cssselect2==0.4.1
dataclasses==0.6
defusedxml==0.6.0
dependency-injector==4.34.0
distlib==0.3.1
filelock==3.0.12
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.2.0
Flask-JWT-Extended==3.25.0
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mobility==1.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-SeaSurf==0.2.2
Flask-SQLAlchemy==2.4.4
flask-talisman==0.7.0
gunicorn==20.0.4
html5lib==1.1
icalendar==4.0.7
identify==1.5.12
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
limits==1.5.1
login==0.0.6
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mistune==0.8.4
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==21.0
parameterized==0.8.1
passlib==1.7.4
pathspec==0.8.1
pdfkit==0.6.1
Pillow==8.1.0
platformdirs==2.0.2
pluggy==0.13.1
pre-commit==2.9.3
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
Pyphen==0.10.0
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-pdf==0.37
pytz==2020.4
PyYAML>=5.4
redis==3.5.3
regex==2020.11.13
requests==2.25.1
rq==1.7.0
rq-dashboard==0.6.1
six==1.15.0
SQLAlchemy==1.3.21
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.3.1
WeasyPrint==52.2
webencodings==0.5.1
Werkzeug==1.0.1
